Moved dao and entities to another package
