+++ app/src/main/java/com/patrycja/filip/machinelearning/persistence/db/entity/ChapterProgressEntity.java
@@ -1,11 +1,9 @@
-package com.patrycja.filip.machinelearning.db.entity;
+package com.patrycja.filip.machinelearning.persistence.db.entity;
 
 import android.arch.persistence.room.ColumnInfo;
 import android.arch.persistence.room.Entity;
 import android.arch.persistence.room.ForeignKey;
 import android.arch.persistence.room.PrimaryKey;
-
-import java.util.Date;
 
 import lombok.Data;
